% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{add_message.GPT_messages}
\alias{add_message.GPT_messages}
\title{Add a message to a GPT_messages object}
\usage{
\method{add_message}{GPT_messages}(x, content, role = "user")
}
\description{
Add a message to a GPT_messages object
}
