% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpaca_inclusion_criteria}
\alias{alpaca_inclusion_criteria}
\title{Example inclusion criteria for source screening}
\format{
A list with four elements
\describe{
\item{study_objective}{The objective of the review, a character vector of length 1}
\item{population}{The review population, a character vector of length 1}
\item{context}{The review context, a character vector of length 1}
\item{concept}{The review concept, a character vector of length 1}
}
}
\usage{
alpaca_inclusion_criteria
}
\description{
Inclusion criteria for a hypothetical scoping review
}
\seealso{
alpaca_sources
}
\keyword{datasets}
