% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpaca_results}
\alias{alpaca_results}
\title{Example results of source screening}
\format{
A data frame with ten rows and five columns:
\describe{
\item{title}{The source title}
\item{abstract}{The source abstract}
\item{include}{Whether the source meets the description given in \code{alpaca_review_description}}
\item{GPT_conversation}{The conversation with GPT in which it is lead
through considering the source and making a recommendation to include or
exclude, a list-column of \code{GPT_mssg} objects}
\item{GPT_recommendation}{GPT's recommendation, either the word 'INCLUDE'
or 'EXCLUDE'}
}
}
\usage{
alpaca_results
}
\description{
A data frame of example results from screening sources with the \code{screen_sources()} function.
}
\seealso{
alpaca_review_description alpaca_sources
}
\keyword{datasets}
