% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpaca_review_description}
\alias{alpaca_review_description}
\title{Example review description for source screening}
\format{
A list with four elements
\describe{
\item{objective}{The objective of the review, a character vector of length 1}
\item{population}{The review population, a character vector of length 1}
\item{context}{The review context, a character vector of length 1}
\item{concept}{The review concept, a character vector of length 1}
}
}
\usage{
alpaca_review_description
}
\description{
Description of a hypothetical scoping review.
}
\seealso{
alpaca_sources alpaca_results
}
\keyword{datasets}
