% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpaca_sources}
\alias{alpaca_sources}
\title{Example sources for screening}
\format{
A list with ten elements, each contaning:
\describe{
\item{title}{The source title, a character vector of length 1}
\item{abstract}{The source abstract, a character vector of length 1}
\item{include}{Whether the source meets the inclusion criteria given in \code{alpaca_inclusion_criteria}, a logical vector of length 1}
}
}
\usage{
alpaca_sources
}
\description{
Sources to be screened for a hypothetical scoping review
}
\seealso{
alpaca_inclusion_criteria
}
\keyword{datasets}
