% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{as.data.frame.GPT_messages}
\alias{as.data.frame.GPT_messages}
\title{Convert GPT_messages object to data frame}
\usage{
\method{as.data.frame}{GPT_messages}(x, ...)
}
\arguments{
\item{x}{A GPT_messages object}

\item{...}{Other arguments to be bassed to data.frame()}
}
\description{
Convert GPT_messages object to data frame
}
