% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{as.vector.GPT_messages}
\alias{as.vector.GPT_messages}
\title{Convert GPT_messages object to a human-readable vector}
\usage{
\method{as.vector}{GPT_messages}(x, mode = "any")
}
\description{
Convert GPT_messages object to a human-readable vector
}
