% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{check_API_key}
\alias{check_API_key}
\title{Check that the OPENAI_API_KEY environmental variable is set}
\usage{
check_API_key(show_key = FALSE)
}
\arguments{
\item{show_key}{If TRUE, and if the environmental variable is set, the key
will be printed to the console. Defaults to FALSE.}
}
\description{
Check that the OPENAI_API_KEY environmental variable is set
}
