% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{complete_GPT.GPT_messages}
\alias{complete_GPT.GPT_messages}
\title{Create chat completions with the OpenAI API}
\usage{
\method{complete_GPT}{GPT_messages}(x, .dry_run = FALSE)
}
\arguments{
\item{x}{A GPT_messages object}

\item{.dry_run}{If TRUE, will not make an actual call to the GPT API}
}
\description{
API endpoint is documented at
\url{https://platform.openai.com/docs/api-reference/chat/create}
}
