% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{complete_GPT_tryCatch}
\alias{complete_GPT_tryCatch}
\title{tryCatch-buffered complete_GPT}
\usage{
complete_GPT_tryCatch(x, tries = 3, .dry_run = FALSE)
}
\arguments{
\item{x}{A GPT_messages object}

\item{tries}{How many times to try before giving up}

\item{.dry_run}{If TRUE, will not make an actual call to the GPT API}
}
\description{
tryCatch-buffered complete_GPT
}
