% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{create_chat_completion}
\alias{create_chat_completion}
\title{Create chat completions with the OpenAI API}
\usage{
create_chat_completion(messages)
}
\description{
Create chat completions with the OpenAI API
}
