% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{GPT_messages}
\alias{GPT_messages}
\title{Helper to create GPT_messages objects}
\usage{
GPT_messages(role = character(), content = character())
}
\description{
Helper to create GPT_messages objects
}
