% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPTscreenR-package.R
\docType{package}
\name{GPTscreenR-package}
\alias{GPTscreenR}
\alias{GPTscreenR-package}
\title{GPTscreenR: GPT-assisted source screening for scoping reviews}
\description{
GPT-assisted source screening for scoping reviews.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wilkox/GPTscreenR}
  \item Report bugs at \url{https://github.com/wilkox/GPTscreenR/issues}
}

}
\author{
\strong{Maintainer}: David Wilkins \email{david@wilkox.org}

}
\keyword{internal}
