% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{last_message.GPT_messages}
\alias{last_message.GPT_messages}
\title{Return the last message from a GPT_messages object}
\usage{
\method{last_message}{GPT_messages}(x)
}
\arguments{
\item{x}{A GPT_messages object}
}
\description{
Return the last message from a GPT_messages object
}
