% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{new_GPT_messages}
\alias{new_GPT_messages}
\title{A constructor function for GPT_messages}
\usage{
new_GPT_messages(x = list(role = character(), content = character()))
}
\description{
A constructor function for GPT_messages
}
