% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{print.GPT_messages}
\alias{print.GPT_messages}
\title{A print method for GPT_messages}
\usage{
\method{print}{GPT_messages}(x, ...)
}
\description{
Coloured output based on
\url{https://github.com/r-lib/testthat/blob/717b02164def5c1f027d3a20b889dae35428b6d7/R/colour-text.r}
}
