% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{review_description}
\alias{review_description}
\title{Combine the review objective and inclusion criteria into a formatted string}
\usage{
review_description(
  objective = NULL,
  population = NULL,
  concept = NULL,
  context = NULL
)
}
\arguments{
\item{objective}{A brief description of the overall review objective, a
character vector of length 1}

\item{population}{A brief description of the overall review objective, a
character vector of length 1}

\item{concept}{A brief description of the overall review objective, a
character vector of length 1}

\item{context}{A brief description of the overall review objective, a
character vector of length 1}
}
\description{
Combine the review objective and inclusion criteria into a formatted string
}
