% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_source}
\alias{screen_source}
\title{Screen a source against study inclusion criteria}
\usage{
screen_source(
  study_objective,
  population,
  concept,
  context,
  title,
  abstract,
  .verbose = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{study_objective}{A brief description of the overall study objective, a
character vector of length 1}

\item{population}{A brief description of the overall study objective, a
character vector of length 1}

\item{concept}{A brief description of the overall study objective, a
character vector of length 1}

\item{context}{A brief description of the overall study objective, a
character vector of length 1}

\item{title}{The title of the article to be screened, a character vector of
length 1}

\item{abstract}{The abstract of the article to be screened, a character
vector of length 1}

\item{.verbose}{If FALSE, progress messages will be suppressed}

\item{.dry_run}{If TRUE, calls to the GPT API will be skipped}
}
\description{
Screen a source against study inclusion criteria
}
