% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen_sources}
\alias{screen_sources}
\title{Screen multiple sources against study inclusion criteria}
\usage{
screen_sources(
  sources,
  study_description,
  n = NULL,
  random = TRUE,
  cache_file = fs::path("sources_cache.rds"),
  .verbose = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{sources}{A data frame of sources to screen, containing at least
'title' and 'abstract' columns}

\item{study_description}{A description of the study including objective and
inclusion criteria, a character vector of length 1}

\item{n}{Optional, a maximum number of sources to screen}

\item{random}{A logical value indicating whether to randomise the order in
which sources are screened, defaults to TRUE}

\item{cache_file}{A file in which the sources list cache is kept (as RDS),
either fs::path() or character vector of length 1}

\item{.verbose}{If FALSE, progress messages will be suppressed}

\item{.dry_run}{If TRUE, calls to the GPT API will be skipped}
}
\description{
To support multiple screening sessions, a cache of the sources list is
written to a file.
}
