% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{study_description}
\alias{study_description}
\title{Combine study objective and inclusion criteria into a formatted string}
\usage{
study_description(
  objective = NULL,
  population = NULL,
  concept = NULL,
  context = NULL
)
}
\arguments{
\item{objective}{A brief description of the overall study objective, a
character vector of length 1}

\item{population}{A brief description of the overall study objective, a
character vector of length 1}

\item{concept}{A brief description of the overall study objective, a
character vector of length 1}

\item{context}{A brief description of the overall study objective, a
character vector of length 1}
}
\description{
Combine study objective and inclusion criteria into a formatted string
}
