% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{validate_GPT_messages}
\alias{validate_GPT_messages}
\title{A validator function for GPT_messages}
\usage{
validate_GPT_messages(x)
}
\arguments{
\item{x}{A GPT_messages object}
}
\description{
A validator function for GPT_messages
}
