% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpaca_sources}
\alias{alpaca_sources}
\title{Example sources for screening}
\format{
A data frame with ten rows and three columns:
\describe{
\item{title}{The source title}
\item{abstract}{The source abstract}
\item{include}{Whether the source meets the description given in
\code{alpaca_study_description}}
}
}
\usage{
alpaca_sources
}
\description{
A data frame of sources to be screened for a hypothetical scoping review.
}
\seealso{
alpaca_study_description alpaca_results
}
\keyword{datasets}
