% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{as_GPT_messages.data.frame}
\alias{as_GPT_messages.data.frame}
\title{Convert data frame to GPT_messages object}
\usage{
\method{as_GPT_messages}{data.frame}(df = data.frame())
}
\arguments{
\item{df}{A data frame with columns 'role' and 'content' of character type}
}
\description{
Convert data frame to GPT_messages object
}
