% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{GPT_messages}
\alias{GPT_messages}
\title{Helper to create GPT_messages objects}
\usage{
GPT_messages(role = character(), content = character())
}
\arguments{
\item{role}{A character vector of roles in the conversation ("system", "user" or "assistant")}

\item{content}{A character vector of messages in the conversation}
}
\description{
Helper to create GPT_messages objects
}
