% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPT_messages.R
\name{remove_message.GPT_messages}
\alias{remove_message.GPT_messages}
\title{Remove the last message from a GPT_messages object}
\usage{
\method{remove_message}{GPT_messages}(x)
}
\arguments{
\item{x}{A GPT_messages object}
}
\description{
Remove the last message from a GPT_messages object
}
